backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
sentence-transformers